requirements.txt
djangorestframework-jsonapi==4.2.0
